+++ app/src/lib/menu-item.ts
@@ -1,0 +1,13 @@
+export interface IMenuItem {
+  /** The user-facing label. */
+  readonly label?: string
+
+  /** The action to invoke when the user selects the item. */
+  readonly action?: () => void
+
+  /** The type of item. */
+  readonly type?: 'separator'
+
+  /** Is the menu item enabled? Defaults to true. */
+  readonly enabled?: boolean
+}
